mv kallisto index fixture to test_kallisto
